backend/requirements.txt
fastapi
uvicorn[standard]
requests
google-generativeai>=0.3.0
python-jose
